backend/superDeskAgent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
aws-requests-auth==0.4.3
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.51
botocore==1.40.51
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dill==0.4.0
docstring_parser==0.17.0
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.8.0
google-auth==2.41.1
google-genai==1.43.0
h11==0.16.0
halo==0.0.31
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
log-symbols==0.0.14
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.3
mcp==1.17.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
numpy==2.2.6
ollama==0.6.0
opencv-python==4.12.0.88
opentelemetry-api==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
slack_bolt==1.26.0
slack_sdk==3.37.0
sniffio==1.3.1
soupsieve==2.8
spinners==0.0.24
sse-starlette==3.0.2
starlette==0.48.0
strands-agents==1.12.0
strands-agents-builder==0.1.10
strands-agents-tools==0.2.11
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
flask
flask-cors
flask-socketio
eventlet
python-dotenv
requests


